// types/car.ts
export interface Car {
  id: number;
  name: string;
  slug: string;
  brand: string;
  type: string;
  year: string;
  image: string;

  range: string;
  power: string;
  acceleration: string;
  battery: string;
  charging: string;
  topSpeed: string;

  price: string;

  description: string;

  features: string[];

  colors: string[];
}